# docs/swagger.yaml
basePath: /
definitions:
  controllers.Razor:
    properties:
      OrderID:
        type: string
      PaymentID:
        type: string
      Signature:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: B Y E C O M  is your ecom. shop solution, where you can sell anything
    online with ease and at an affordable price. We provide an eCommerce platform
    for businesses to sell their products online and connect with customers worldwide.
  title: B Y E C O M  LTD
  version: "1.0"
paths:
...
  /user/password:
    patch:
      description: Updating user password
      parameters:
      - description: address id
        in: formData
        name: address
        required: true
        type: string
      - description: coupon code
        in: formData
        name: coupon
        required: true
        type: string
      - description: payment method
        in: formData
        name: method
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Updating Password
      tags:
      - User Profile
  /user/payment:
    get:
      description: Razor Pay payment gateway
      parameters:
      - description: Payment id
        in: query
        name: id
        required: true
        type: string
      responses: {}
      summary: Razor Pay
      tags:
      - User Payment
    post:
      consumes:
      - application/json
      description: Verifying Payment and updating payment status
      parameters:
      - description: Payment details
        in: body
        name: verify
        required: true
        schema:
          $ref: '#/definitions/controllers.Razor'
      produces:
      - application/json
      responses: {}
      summary: Verify Payment
      tags:
      - User Payment
...
  /user/profile:
    get:
      description: Showing User Profile
      responses: {}
      summary: Show Profile
      tags:
      - User Profile
    patch:
      description: Editing user profile
      parameters:
      - description: user's name
        in: formData
        name: name
        required: true
        type: string
      - description: user's phone
        in: formData
        name: phone
        required: true
        type: string
      - description: user's gender
        in: formData
        name: gender
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Editing Profile
      tags:
      - User Profile
  /user/rating:
    post:
      consumes:
      - multipart/form-data
      description: Adding Rating and Review for product
      parameters:
      - description: Product Id
        in: query
        name: id
        required: true
        type: string
      - description: Product review
        in: formData
        name: rating
        required: true
        type: string
      - description: Product rating
        in: formData
        name: review
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Rating Add
      tags:
      - User Rating
    put:
      consumes:
      - multipart/form-data
      description: Editing Rating and Review for product
      parameters:
      - description: Product review
        in: formData
        name: rating
        required: true
        type: string
      - description: Product rating
        in: formData
        name: review
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Rating Editing
      tags:
      - User Rating
  /user/reset:
    post:
      consumes:
      - multipart/form-data
      description: Changing user password without login
      parameters:
      - description: Reset code
        in: query
        name: code
        required: true
        type: string
      - description: New Pass
        in: formData
        name: newpass
        required: true
        type: string
      - description: Repeat Pass
        in: formData
        name: repeatpass
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Changing Pass
      tags:
      - User ForgetPass
  /user/reset/password:
    post:
      consumes:
      - multipart/form-data
      description: Initialilzing forgot password without login
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Forgot Pass Init
      tags:
      - User ForgetPass
...
swagger: "2.0"